# openapi.yaml
openapi: 3.1.0
info:
  title: ratelimit
  version: "1.0.0"
  description: >-
    Central rate limits and quotas for every service in the atc-web platform: named policies with
    several sliding windows, per-subject overrides and hard blocks, one atomic check per request or
    per group, usage and decision statistics. SQLite, no external store. Consumed primarily by
    gateway and other backend services, gated only by API-key possession (no separate network
    tier).


    **Read this before integrating `POST /v1/check` or `POST /v1/check/batch`**: a rate-limit
    *denial* is a normal, successful HTTP response. These two endpoints (and `POST /v1/release`)
    always return HTTP `200` for a well-formed, authorized request — whether the business decision
    is `allowed: true` or `allowed: false` is carried entirely in the JSON body's `allowed` field
    (and, for a batch, in each result's own `allowed` field plus the top-level `allowed`). An HTTP
    `429` from this service means something completely different: it is `ratelimit`'s own
    self-throttle (`@fastify/rate-limit`, keyed by API-key id, scoped to the whole `/v1` sub-app)
    protecting the service itself from being called too often — it has nothing to do with any
    policy's own limits. Do not branch on HTTP status to detect a rate-limit denial; branch on the
    response body's `allowed` field. See each operation below for the exact per-endpoint wording,
...
security:
  - ApiKeyAuth: []
paths:
...
  /v1/check:
    post:
      operationId: ratelimit.check
      summary: Check and consume one subject's quota against one policy
      description: >-
        Evaluates every window of the effective limit set (the subject's active override, if any,
        else the policy's own limits) and, unless `peek: true`, atomically consumes `cost` units
        (default 1) when every window allows it — a denied check never consumes anything.


        **HTTP `200` is returned for both an allow and a deny.** The business decision is the body's
        `allowed` field, never the HTTP status. This is deliberate, not an oversight: only this
        service's own self-throttle (see `components.responses.RateLimited`) ever produces an HTTP
        `429` here, and it is a completely separate mechanism from a policy's own limit — a `429`
        means "you called `ratelimit` itself too often," never "this subject is over its policy
        limit." No `Retry-After` HTTP header is ever set for a business-level denial; the
        domain-level `retryAfter` on the returned `Decision` is body-only.


        Requires a `check` (or `readwrite`) API key, scoped to the target policy (a key with a
        `policies` allowlist that does not include `policy` gets `403 FORBIDDEN`, checked before the
        policy is even looked up).
      x-required-role: check
      x-policy-scoped: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckRequest'
      responses:
        '200':
          description: >-
            The check was evaluated — `allowed` may be `true` or `false`; both are a normal `200`,
            never an HTTP `429`. `blocked: true` means the effective limits include a `limit: 0`
            window, i.e. this subject is hard-blocked regardless of remaining budget elsewhere.
          headers:
            X-Content-Type-Options:
              $ref: '#/components/headers/XContentTypeOptions'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Decision'
        '400':
          description: Validation failed, or the cost could never fit one of the effective windows.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/PolicyNotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/check/batch:
    post:
      operationId: ratelimit.checkBatch
      summary: Check and consume several (policy, subject) pairs as one atomic unit
      description: >-
        Same all-or-nothing atomicity and the same `200`-even-when-denied contract as `POST
        /v1/check` (see that operation's description — it applies identically here): the top-level
        `allowed` and every `results[i].allowed` are body-level business decisions, never reflected
        in the HTTP status, and never produce an HTTP `429` by themselves. Counters move only when
        **every** check in the batch allows its cost; if any one is denied, every result in the
        response comes back with `allowed: false` (the ones that individually would have passed are
        still reported as not-allowed, since nothing was consumed for them either).


        Requires a `check` (or `readwrite`) API key, scoped to every policy named in `checks` (the
        first out-of-scope policy found gets `403 FORBIDDEN`).


        `400 DUPLICATE_CHECK` is a genuinely overloaded error code with two unrelated trigger
        conditions — do not assume it always means a literal duplicate: (1) the batch has more than
        `MAX_BATCH` entries (config-configurable, default 20 — note the request body schema itself
        allows up to 100 entries, so a batch that passes JSON-Schema validation can still be
        rejected here), or (2) the same `(policy, subject)` pair appears twice in `checks`.
      x-required-role: check
      x-policy-scoped: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckBatchRequest'
      responses:
        '200':
          description: >-
            Every check in the batch was evaluated. `allowed` (top-level) is true only if every
            entry in `results` was allowed; a `200` is returned regardless, never a `429`.
          headers:
            X-Content-Type-Options:
              $ref: '#/components/headers/XContentTypeOptions'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckBatchResponse'
        '400':
          description: >-
            Validation failed, a cost could never fit one of the effective windows
            (`COST_TOO_HIGH`), or `DUPLICATE_CHECK` — see this operation's description for its two
            distinct trigger conditions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/PolicyNotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/release:
    post:
      operationId: ratelimit.release
      summary: Give back units consumed by an earlier check
      description: >-
        For an operation that consumed quota via `POST /v1/check` and then itself failed, releases
        `cost` units (default 1) back into the window they were actually consumed from.
        `consumedAt`, when given, should be the `consumedAt` value the original `Decision` returned —
        this targets the window that was truly consumed even after it has rolled over; omitting it
        targets whichever window `now` falls in, which is only correct when the release happens
        inside the same window the original check consumed. Evaluated at `now`, and — like `/check`
        and `/check/batch` — always `200`, with the resulting `Decision` reflecting the state after
        the release; there is no failure mode here that maps to a body-level `allowed: false`
        distinct from the returned decision's own fields.


        Requires a `check` (or `readwrite`) API key, scoped to the target policy.
      x-required-role: check
      x-policy-scoped: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReleaseRequest'
      responses:
        '200':
          description: State after the release, evaluated at `now`.
          headers:
            X-Content-Type-Options:
              $ref: '#/components/headers/XContentTypeOptions'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Decision'
        '400':
          description: Validation failed, cost out of range, or `consumedAt` is not a valid ISO 8601 date.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/PolicyNotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/policies:
    post:
      operationId: ratelimit.policies.create
      summary: Create a policy
      description: >-
        Requires a `write` (or `readwrite`) API key, scoped to the new policy's own name. Returns
        `Location: /v1/policies/{name}` on success.
      x-required-role: write
      x-policy-scoped: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePolicyRequest'
      responses:
        '201':
          description: Policy created.
          headers:
            Location:
              $ref: '#/components/headers/Location'
            X-Content-Type-Options:
              $ref: '#/components/headers/XContentTypeOptions'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyResponse'
        '400':
          description: Validation failed, or `limits` is invalid (see `INVALID_LIMITS`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/PolicyExists'
        '429':
          $ref: '#/components/responses/RateLimited'
    get:
      operationId: ratelimit.policies.list
      summary: List policies visible to this API key
      description: >-
        Requires a `read` (or `readwrite`) API key. Filtered to the key's own `policies` scope when
        it has one (`policies: null` on the key sees every policy). Never scope-forbidden — an
        out-of-scope policy is silently omitted rather than causing a `403`.
      x-required-role: read
      responses:
        '200':
          description: Policies visible to this key.
          headers:
            X-Content-Type-Options:
              $ref: '#/components/headers/XContentTypeOptions'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoliciesListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/policies/{name}:
    get:
      operationId: ratelimit.policies.get
      summary: Get one policy
      description: Requires a `read` (or `readwrite`) API key, scoped to this policy.
      x-required-role: read
      x-policy-scoped: true
      parameters:
        - $ref: '#/components/parameters/PolicyName'
      responses:
        '200':
          description: The policy.
          headers:
            X-Content-Type-Options:
              $ref: '#/components/headers/XContentTypeOptions'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyResponse'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/PolicyNotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
    patch:
      operationId: ratelimit.policies.patch
      summary: Update a policy's description and/or limits
      description: >-
        Requires a `write` (or `readwrite`) API key, scoped to this policy. At least one of
        `description`/`limits` must be given. Existing overrides and counters are unaffected by a
        limits change; only future checks are evaluated against the new limits.
      x-required-role: write
      x-policy-scoped: true
      parameters:
        - $ref: '#/components/parameters/PolicyName'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchPolicyRequest'
      responses:
        '200':
          description: The updated policy.
          headers:
            X-Content-Type-Options:
              $ref: '#/components/headers/XContentTypeOptions'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyResponse'
        '400':
          description: Validation failed (including an empty patch body), or `limits` is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/PolicyNotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
    delete:
      operationId: ratelimit.policies.delete
      summary: Delete a policy
      description: >-
        Requires a `write` (or `readwrite`) API key, scoped to this policy. Deletes the policy along
        with its overrides, counters and statistics. No response body.
      x-required-role: write
      x-policy-scoped: true
      parameters:
        - $ref: '#/components/parameters/PolicyName'
      responses:
        '204':
          description: Policy deleted.
          headers:
            X-Content-Type-Options:
              $ref: '#/components/headers/XContentTypeOptions'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/PolicyNotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/policies/{name}/stats:
    get:
      operationId: ratelimit.policies.getStats
      summary: Hourly allowed/denied series for a policy
      description: >-
        Requires a `read` (or `readwrite`) API key, scoped to this policy. Hourly series for the
        last `hours` hours (default 24, max 720 = 30 days), gaps filled with zeros.
      x-required-role: read
      x-policy-scoped: true
      parameters:
        - $ref: '#/components/parameters/PolicyName'
        - name: hours
          in: query
          required: false
          description: Number of trailing hours to report, 1-720. Default 24.
          schema:
            type: string
            pattern: '^([1-9]|[1-9][0-9]|[1-6][0-9][0-9]|7[0-1][0-9]|720)$'
            example: "24"
      responses:
        '200':
          description: Hourly allowed/denied series.
          headers:
            X-Content-Type-Options:
              $ref: '#/components/headers/XContentTypeOptions'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsResponse'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/PolicyNotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/policies/{name}/top:
    get:
      operationId: ratelimit.policies.getTop
      summary: Subjects with the highest usage in one window of a policy
      description: >-
        Requires a `read` (or `readwrite`) API key, scoped to this policy. `window` defaults to the
        policy's first (shortest) configured window when omitted; a `window` value that is not one
        of the policy's configured windows is `400 UNKNOWN_WINDOW`.
      x-required-role: read
      x-policy-scoped: true
      parameters:
        - $ref: '#/components/parameters/PolicyName'
        - name: window
          in: query
          required: false
          description: Window size in seconds, must match one of the policy's configured windows. Defaults to the policy's shortest window.
          schema:
            type: string
            pattern: '^[1-9][0-9]{0,7}$'
        - name: limit
          in: query
          required: false
          description: Maximum subjects to return, 1-1000. Default 20.
          schema:
            type: string
            pattern: '^([1-9]|[1-9][0-9]|[1-9][0-9][0-9]|1000)$'
      responses:
        '200':
          description: Top subjects by usage in the resolved window.
          headers:
            X-Content-Type-Options:
              $ref: '#/components/headers/XContentTypeOptions'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopResponse'
        '400':
          description: Validation failed, or `window` does not match any configured window (`UNKNOWN_WINDOW`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/PolicyNotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/policies/{name}/overrides:
    get:
      operationId: ratelimit.overrides.list
      summary: List a policy's subject overrides
      description: Requires a `read` (or `readwrite`) API key, scoped to this policy. Offset-paginated.
      x-required-role: read
      x-policy-scoped: true
      parameters:
        - $ref: '#/components/parameters/PolicyName'
        - name: limit
          in: query
          required: false
          description: Page size, 1-1000. Default 50.
          schema:
            type: string
            pattern: '^([1-9]|[1-9][0-9]|[1-9][0-9][0-9]|1000)$'
        - name: offset
          in: query
          required: false
          description: Rows to skip. Default 0.
          schema:
            type: string
            pattern: '^(0|[1-9][0-9]{0,6})$'
      responses:
        '200':
          description: Overrides for this policy.
          headers:
            X-Content-Type-Options:
              $ref: '#/components/headers/XContentTypeOptions'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OverridesListResponse'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/PolicyNotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /v1/policies/{name}/subjects/{subject}:
    get:
      operationId: ratelimit.subjects.getUsage
      summary: Current usage of one subject, without consuming
      description: >-
        Requires a `read` (or `readwrite`) API key, scoped to this policy. A peek: `remaining` in
        the returned `usage.limits[]` is what is left right now, and no counters are touched. Also
        returns the subject's active override, if any.
      x-required-role: read
      x-policy-scoped: true
      parameters:
        - $ref: '#/components/parameters/PolicyName'
        - $ref: '#/components/parameters/Subject'
      responses:
        '200':
          description: Current usage and, if present, the active override.
          headers:
            X-Content-Type-Options:
              $ref: '#/components/headers/XContentTypeOptions'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectUsageResponse'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/PolicyNotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /v1/stats:
    get:
      operationId: ratelimit.stats.get
      summary: Aggregate statistics across every policy visible to this API key
      description: >-
        Requires a `read` (or `readwrite`) API key. Scoped the same way as `GET /v1/policies` — only
        policies within the key's own `policies` allowlist (or every policy, when unscoped)
        contribute to the totals and `items`.
      x-required-role: read
      responses:
        '200':
          description: Aggregate statistics.
          headers:
            X-Content-Type-Options:
              $ref: '#/components/headers/XContentTypeOptions'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GlobalStatsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: >-
        `Authorization: Bearer <secret>`. Keys come from the `RATELIMIT_API_KEYS` environment
        variable (`id:secret[:role[:policy+policy]]`), compared in constant time against every
        configured key (no early exit on match). Each key carries:


        - a **role** — `check`, `read`, `write` or `readwrite` (default when omitted). A route
          declares the role it needs (see each operation's `x-required-role`); a key satisfies a
          requirement of `need` when its own role is either `need` itself or `readwrite` —
          `readwrite` is the only role that satisfies every requirement, `check`/`read`/`write` each
          satisfy only themselves. There is no partial overlap (e.g. `write` does NOT imply `read`).
          `POST /v1/check`, `POST /v1/check/batch` and `POST /v1/release` specifically require
          `check` (or `readwrite`) — a `read`- or `write`-only key gets `403 FORBIDDEN` on those
          three routes even though it may be valid for other routes.
        - an optional **policy scope** — `policies: string[]|null` (`null` = every policy, the
          default when no policy list is given in the env value). Every route that names a policy
          (by path parameter or by a `policy` field in the request body) additionally requires the
          key's scope to include that policy name, on top of satisfying the role check above;
          out-of-scope also yields `403 FORBIDDEN` (message text distinguishes "no `<role>` access"
          from "no access to policy `<name>`", but the HTTP status and error `code` are identical).


        401 body is always `{"error":{"code":"UNAUTHORIZED","message":"missing or invalid API key"}}`
        with a `WWW-Authenticate: Bearer` header; 403 is always
        `{"error":{"code":"FORBIDDEN","message":"..."}}`.
...